+++ lib/routes/routes.dart
@@ -3,4 +3,7 @@ class Routes {
 
   //Home
   static const COURSE_SCREEN = "/course_screen";
+  static const TOPIC_SCREEN = "/topic_screen";
+  static const MATERIAL_SCREEN = "/material_screen";
+
 }
